+++ subbrute_protocols.c
@@ -612,7 +612,7 @@ void subbrute_protocol_default_payload(
     SubBruteFileProtocol file,
     uint64_t step,
     uint8_t bits,
-    uint8_t te,
+    uint32_t te,
     uint8_t repeat) {
     FuriString* candidate = furi_string_alloc();
     subbrute_protocol_create_candidate_for_default(candidate, file, step);
@@ -647,7 +647,7 @@ void subbrute_protocol_file_payload(
     Stream* stream,
     uint64_t step,
     uint8_t bits,
-    uint8_t te,
+    uint32_t te,
     uint8_t repeat,
     uint8_t bit_index,
     uint64_t file_key,
@@ -690,7 +690,7 @@ void subbrute_protocol_default_generate_file(
     SubBruteFileProtocol file,
     uint64_t step,
     uint8_t bits,
-    uint8_t te,
+    uint32_t te,
     uint8_t repeat) {
     FuriString* candidate = furi_string_alloc();
     subbrute_protocol_create_candidate_for_default(candidate, file, step);
@@ -733,7 +733,7 @@ void subbrute_protocol_file_generate_file(
     SubBruteFileProtocol file,
     uint64_t step,
     uint8_t bits,
-    uint8_t te,
+    uint32_t te,
     uint8_t repeat,
     uint8_t bit_index,
     uint64_t file_key,
